requirements.txt
-e .
numpy==1.18.3
scipy==1.4.1
Pillow==7.0.0
seaborn==0.10.1
Wand==0.6.1
opencv-python==4.2.0.34
scikit-image==0.17.2
pytest==5.4.1
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
Jinja2==2.10.1
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
notebook==6.0.3
pandas==1.0.3
matplotlib==3.2.1
tensorboard==2.2.1
mypy==0.770
torch-summary==1.2.0
-e git://github.com/camgbus/pytorch_template.git@4b3ab702c7308760beb5e559393bd636776ff37f#egg=ptt